+++ src/TreeBuilder/InTableText.php
@@ -52,6 +52,14 @@ class InTableText extends InsertionMode {
 	}
 
 	/**
+	 * Flush the pending table text and restore the mode
+	 */
+	public function flush() {
+		$this->processPendingCharacters();
+		$this->dispatcher->restoreMode();
+	}
+
+	/**
 	 * Common code for the "anything else" case. Process the pending table
 	 * character tokens.
 	 */